Move utils for account filter to a separate file
